Update overlaps.R
I have conducted thorough tests and successfully build a function which can create with no limits. But noticed that when the number of genes exceeds 10k, this function may cause excessive memory usage and might fail to export the graph. 
In addition, for security considerations, a "for" loop was not used to construct !!sym -like functions in rlang package. Therefore, currently, only 2 to 5 networks are supported for overlap.